refactor get-set graph computation out of ssa pass
